Allow to build Arch Linux package with Docker
This adds initial versions of `PKGBUILD` and `Dockerfile` that allow one
to build an Arch Linux package with `logdl` like this: `docker build -t
logdl-build .`. The resulting package is located in the built container
(this should be updated to extract it on the host).

Prerequisite: the linux binary should already be built in Docker using
`stack --docker build`.