Don't do anything if there are no matching log files
This fixes a meaningless prompt "Remove ?" when the user wanted to
remove log files but there were none; the download action was already
silent in this case.

This is implemented by converting the list of `RemoteFile`s into a
`NonEmpty RemoteFile` (although it's converted back to a list after
that) and the result is wrapped in a `Maybe`. So the solution is to add
the `MaybeT` transformer in the `run`'s `do` block — this allows the
code to use the `Config`, `IO` actions, and also stop executing if it
gets a `Nothing` (for an empty list of files).

This is my second implementation attempt. The first one started the same
by adding `MaybeT` but I couldn't wrap my head around the constant type
mismatch errors. The main issue there was the `lift` functions when
binding `tags` and `today`, replacing them with `liftIO` calms down the
compiler significantly — must be because `lift` lifts a value one level
of the transformer stack, which was fine before (`IO` -> `ReaderT Config
IO`) but not now (`IO` -> `MaybeT (ReaderT Config IO)`), and `liftIO`
lifts the `IO` actions to the right level (always the last one).