Get rid of early termination
This early `exitSuccess` would mess up the flow of pre- and post-actions
(like waiting for the server to appear and disappear, will be added
soon), so the delete `actionF` is muted if the user doesn't agree to
remove files. Not the most efficient implementation because the later
code `traverse`s over available files and calls action on every one of
them, but that's a very minor issue.