Fix a bug when waiting for disappear didn't work if there are no files
This happened because `run` was in the `MaybeT (ReaderT Config IO)`
monad, so if the extracted `files` array was empty, it would be
transformed into `Nothing`, stopping the execution. The fix is to remove
`MaybeT` from the `run` function directly and use it only while handling
an action — it's still useful there because we don't want to ask to
confirm the deletion of no files.